watch keyspace level gc
