+++ src/server/gc_worker/gc_manager.rs
@@ -10,13 +10,19 @@ use std::{
     time::Duration,
 };
 use std::sync::atomic::AtomicI64;
+use futures::{compat::Future01CompatExt, stream, StreamExt};
 use dashmap::DashMap;
+use kvproto::meta_storagepb::EventEventType;
 
 use engine_traits::KvEngine;
 use kvproto::metapb::Region;
-use pd_client::FeatureGate;
+use pd_client::{FeatureGate, RpcClient};
+use pd_client::{PdClient,Error as PdError};
+use pd_client::meta_storage::{MetaStorageClient,Checked, Get, Sourced, Watch};
 use raftstore::coprocessor::RegionInfoProvider;
 use tikv_util::{store::find_peer, time::Instant, worker::Scheduler};
+use tikv_util::timer::GLOBAL_TIMER_HANDLE;
+use tikv_util::worker::Worker;
 use txn_types::{Key, TimeStamp};
 
 use super::{
@@ -33,6 +39,8 @@ const BEGIN_KEY: &[u8] = b"";
 
 const PROCESS_TYPE_GC: &str = "gc";
 const PROCESS_TYPE_SCAN: &str = "scan";
+
+const RETRY_INTERVAL: Duration = Duration::from_secs(1); // to consistent with pd_client
 
 /// The configurations of automatic GC.
 pub struct AutoGcConfig<S: GcSafePointProvider, R: RegionInfoProvider> {
@@ -250,7 +258,6 @@ pub(super) struct GcManager<S: GcSafePointProvider, R: RegionInfoProvider, E: Kv
 
     max_concurrent_tasks: usize,
 
-    ks_safepoint_v2: Arc<DashMap<u32, u64>>,
 
 }
 
@@ -272,7 +279,6 @@ impl<S: GcSafePointProvider, R: RegionInfoProvider + 'static, E: KvEngine> GcMan
             cfg_tracker,
             feature_gate,
             max_concurrent_tasks: concurrent_tasks,
-            ks_safepoint_v2:Arc::new(DashMap::default()),
         }
     }
 
@@ -667,6 +673,168 @@ impl<S: GcSafePointProvider, R: RegionInfoProvider + 'static, E: KvEngine> GcMan
             }
         }
     }
+}
+
+
+
+#[derive(Clone)]
+pub struct KeyspaceLevelGCWatchService {
+    pd_client: Arc<RpcClient>,
+    // wrap for etcd client.
+    meta_client: Checked<Sourced<Arc<RpcClient>>>,
+    // record watch revision.
+    revision: i64,
+
+    ks_safepoint_v2: Arc<DashMap<u32, u64>>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+struct KeyspaceLevelGC {
+    keyspace_id: u32,
+    safe_point: u64,
+}
+
+impl KeyspaceLevelGCWatchService {
+    /// Constructs a new `Service` with `ResourceGroupManager` and a
+    /// `RpcClient`.
+    pub fn new(
+        pd_client: Arc<RpcClient>,
+        keyspace_level_gc: Arc<DashMap<u32, u64>>,
+    ) -> KeyspaceLevelGCWatchService {
+        KeyspaceLevelGCWatchService {
+            revision: 0,
+            meta_client: Checked::new(Sourced::new(
+                Arc::clone(&pd_client.clone()),
+                pd_client::meta_storage::Source::KeysapceLevelGC,
+            )),
+            pd_client,
+            ks_safepoint_v2:keyspace_level_gc,
+        }
+    }
+
+    pub async fn watch_keyspace_gc(&mut self) {
+        info!("[test-yjy]watch_keyspace_gc");
+        // Firstly, load all resource groups as of now.
+        self.reload_all_keyspace_level_gc().await;
+        let keyspace_level_gc_prefix=self.get_keyspcace_level_gc_prefix();
+
+        'outer: loop {
+            // Secondly, start watcher at loading revision.
+            let (mut stream, cancel) = stream::abortable(
+                self.meta_client.watch(
+                    Watch::of(keyspace_level_gc_prefix.as_str())
+                        .prefixed()
+                        .from_rev(self.revision)
+                        .with_prev_kv(),
+                ),
+            );
+            info!("pd meta client creating watch stream."; "path" => keyspace_level_gc_prefix.clone(), "rev" => %self.revision);
+            while let Some(grpc_response) = stream.next().await {
+                match grpc_response {
+                    Ok(resp) => {
+                        self.revision = resp.get_header().get_revision();
+                        let events = resp.get_events();
+                        events.iter().for_each(|event| match event.get_type() {
+                            EventEventType::Put => {
+                                info!("[test-yjy]EventEventType::Put01");
+                                let val=event.get_kv().get_value();
+                                match serde_json::from_slice::<KeyspaceLevelGC>(
+                                    event.get_kv().get_value(),
+                                ) {
+                                    Ok(keyspace_level_gc) => {
+                                        self.ks_safepoint_v2.insert(keyspace_level_gc.keyspace_id, keyspace_level_gc.safe_point);
+                                        info!("[test-yjy]EventEventType::Put02-01 {},{}",keyspace_level_gc.keyspace_id,keyspace_level_gc.safe_point);
+                                    },
+                                    Err(e) => error!("parse put keyspace level gc event failed";  "err" => ?e),
+                                }
+                            }
+                            EventEventType::Delete => {
+                                info!("[test-yjy]EventEventType::Delete01");
+                                match serde_json::from_slice::<KeyspaceLevelGC>(
+                                    event.get_kv().get_value(),
+                                ) {
+                                    Ok(keyspace_level_gc) => {
+                                        self.ks_safepoint_v2.remove(&keyspace_level_gc.keyspace_id);
+                                        info!("[test-yjy]EventEventType::Put02-01");
+                                    },
+                                    Err(e) => error!("parse delete keyspace level gc event failed"; "name" => ?event.get_kv().get_key(), "err" => ?e),
+                                }
+                            }});
+                    }
+                    Err(PdError::DataCompacted(msg)) => {
+                        error!("required revision has been compacted"; "err" => ?msg);
+                        //self.reload_all_resource_groups().await;
+                        info!("[test-yjy] PdError::DataCompacted");
+                        self.reload_all_keyspace_level_gc().await;
+                        cancel.abort();
+                        continue 'outer;
+                    }
+                    Err(err) => {
+                        error!("failed to watch resource groups"; "err" => ?err);
+                        let _ = GLOBAL_TIMER_HANDLE
+                            .delay(std::time::Instant::now() + RETRY_INTERVAL)
+                            .compat()
+                            .await;
+                        cancel.abort();
+                        continue 'outer;
+                    }
+                }
+            }
+        }
+    }
+
+    fn get_keyspcace_level_gc_prefix(&self) -> String {
+        let cluster_id = self.pd_client
+            .get_cluster_id()
+            .unwrap();
+        let keyspace_level_gc_prefix=format!("/pd/{}/keyspaces/gc_safe_point", cluster_id);
+        return keyspace_level_gc_prefix;
+    }
+
+    async fn reload_all_keyspace_level_gc(&mut self) {
+        let keyspace_level_gc_prefix=self.get_keyspcace_level_gc_prefix();
+        loop {
+            match self
+                .meta_client
+                .get(Get::of(keyspace_level_gc_prefix.as_str()).prefixed())
+                .await
+            {
+                Ok(mut resp) => {
+                    let kvs = resp.take_kvs().into_iter().collect::<Vec<_>>();
+                    kvs.iter().for_each(|g| {
+                        match serde_json::from_slice::<KeyspaceLevelGC>(
+                            g.get_value(),
+                        ) {
+                            Ok(keyspace_level_gc) => {self.ks_safepoint_v2.insert(keyspace_level_gc.keyspace_id, keyspace_level_gc.safe_point);info!("[test-yjy]EventEventType::Put02-01 {},{}",keyspace_level_gc.keyspace_id,keyspace_level_gc.safe_point);},
+                            Err(e) => error!("parse put keyspace level gc event failed"; "name" => ?g.get_key(), "err" => ?e),
+                        }
+                    });
+
+                    self.revision = resp.get_header().get_revision();
+                    return;
+                }
+                Err(err) => {
+                    error!("failed to get meta storage's keyspace level gc"; "err" => ?err);
+                    let _ = GLOBAL_TIMER_HANDLE
+                        .delay(std::time::Instant::now() + RETRY_INTERVAL)
+                        .compat()
+                        .await;
+                }
+            }
+        }
+    }
+}
+
+pub fn start_periodic_gc_tasks(
+    pd_client: Arc<RpcClient>,
+    bg_worker: &Worker,
+    keyspace_level_gc_cache:Arc<DashMap<u32, u64>>,
+) {
+    let mut keyspace_level_gc_watch_service = KeyspaceLevelGCWatchService::new(pd_client,keyspace_level_gc_cache);
+    // spawn a task to watch all resource groups update.
+    bg_worker.spawn_async_task(async move {
+        keyspace_level_gc_watch_service.watch_keyspace_gc().await;
+    });
 }
 
 #[cfg(test)]
